software/control-bridge/requirements.txt
python-osc==1.8.3
pyserial==3.5
PyYAML==6.0.2
